Everything works. Have to implement lazy loading
